Ya se pueden agregar clientes y el modelo para cancelar pedidos está listo

+++ serigrafia/resources/views/admin/pedido/registroPedido.blade.php
@@ -46,8 +46,8 @@
             </div>
         </center>
     </div>
-    <a href="{{route('pedidos.agregarProducto')}}">Registrar producto</a>
-    <a href="{{route('pedidos.agregarCliente')}}">Registrar cliente</a>
+    <a class="mr-8 ml-8" href="{{route('pedidos.agregarProducto')}}">Registrar producto</a>
+    <a class="" href="{{route('pedidos.agregarCliente')}}">Registrar cliente</a>
 </div>
 <form action="{{route('pedidos.save')}}" method="post" enctype="multipart/form-data" class="p-4">
     @csrf
@@ -100,30 +100,52 @@
 <script type="text/javascript">
     function addProduct(){
         var producto = document.getElementById("Producto");
-        $('#productsTableBody').append(
-            "<tr class='flex flex-wrap sm:table-row mb-6 sm:mb-0'>"
-                +"<td class='w-1/2 p-3 text-white bg-blue-900 rounded-tl-lg sm:hidden border border-b-0 border-r-0'>"
-                    +"Producto"
-                +"</td>"
-                +"<td class='w-1/2 sm:w-auto p-3 text-right sm:text-center border border-b-0 sm:border-none rounded-tr-lg sm:rounded-tr-none'>"
-                    +"<input type='hidden' name='productos["+producto.value+"][id]' value='"+producto.value+"'>"
-                    +$('#Producto option:selected').text()
-                +"</td>"
-                +"<td class='w-1/2 p-3 text-white bg-blue-900 rounded-tl-lg sm:hidden border border-b-0 border-r-0'>"
-                    +"Cantidad"
-                +"</td>"
-                +"<td class='w-1/2 sm:w-auto p-3 text-right sm:text-center border border-b-0 sm:border-none rounded-tr-lg sm:rounded-tr-none'>"
-                    +"<input type='number' name='productos["+producto.value+"][cantidad]' value='"+document.getElementById("Cantidad").value+"'>"
-                +"</td>"
-                +"<td class='w-full sm:w-auto p-3 text-center border sm:border-none rounded-b-lg sm:rounded-br-none'>"
-                    +"<button type='button' class='appearance-none bg-pink-600 hover:bg-red-600 text-white transition duration-500 ease-in-out font-bold py-2 px-2 rounded-full focus:outline-none focus:shadow-outline' , onclick='deleteProduct(this)'>"
-                        +"<svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 16 16' width='16' height='16'>"
-                            +"<path fill-rule='evenodd' d='M6.5 1.75a.25.25 0 01.25-.25h2.5a.25.25 0 01.25.25V3h-3V1.75zm4.5 0V3h2.25a.75.75 0 010 1.5H2.75a.75.75 0 010-1.5H5V1.75C5 .784 5.784 0 6.75 0h2.5C10.216 0 11 .784 11 1.75zM4.496 6.675a.75.75 0 10-1.492.15l.66 6.6A1.75 1.75 0 005.405 15h5.19c.9 0 1.652-.681 1.741-1.576l.66-6.6a.75.75 0 00-1.492-.149l-.66 6.6a.25.25 0 01-.249.225h-5.19a.25.25 0 01-.249-.225l-.66-6.6z'></path>"
-                        +"</svg>"
-                    +"</button>"
-                +"</td>"
-            +"</tr>"
-        );
+        var existente = 0
+        $(".idProductos").each(function() 
+        {
+            if (producto.value == this.value) 
+            {
+                existente = 1
+            }
+        });
+        
+        if(existente != 1)
+        {
+            $('#productsTableBody').append(
+                "<tr class='flex flex-wrap sm:table-row mb-6 sm:mb-0'>"
+                    +"<td class='w-1/2 p-3 text-white bg-blue-900 rounded-tl-lg sm:hidden border border-b-0 border-r-0'>"
+                        +"Producto"
+                    +"</td>"
+                    +"<td class='w-1/2 sm:w-auto p-3 text-right sm:text-center border border-b-0 sm:border-none rounded-tr-lg sm:rounded-tr-none'>"
+                        +"<input type='hidden' name='productos["+producto.value+"][id]' value='"+producto.value+"' class='idProductos'>"
+                        +$('#Producto option:selected').text()
+                    +"</td>"
+                    +"<td class='w-1/2 p-3 text-white bg-blue-900 rounded-tl-lg sm:hidden border border-b-0 border-r-0'>"
+                        +"Cantidad"
+                    +"</td>"
+                    +"<td class='w-1/2 sm:w-auto p-3 text-right sm:text-center border border-b-0 sm:border-none rounded-tr-lg sm:rounded-tr-none'>"
+                        +"<input type='number' name='productos["+producto.value+"][cantidad]' value='"+document.getElementById("Cantidad").value+"'>"
+                    +"</td>"
+                    +"<td class='w-full sm:w-auto p-3 text-center border sm:border-none rounded-b-lg sm:rounded-br-none'>"
+                        +"<button type='button' class='appearance-none bg-pink-600 hover:bg-red-600 text-white transition duration-500 ease-in-out font-bold py-2 px-2 rounded-full focus:outline-none focus:shadow-outline' , onclick='deleteProduct(this)'>"
+                            +"<svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 16 16' width='16' height='16'>"
+                                +"<path fill-rule='evenodd' d='M6.5 1.75a.25.25 0 01.25-.25h2.5a.25.25 0 01.25.25V3h-3V1.75zm4.5 0V3h2.25a.75.75 0 010 1.5H2.75a.75.75 0 010-1.5H5V1.75C5 .784 5.784 0 6.75 0h2.5C10.216 0 11 .784 11 1.75zM4.496 6.675a.75.75 0 10-1.492.15l.66 6.6A1.75 1.75 0 005.405 15h5.19c.9 0 1.652-.681 1.741-1.576l.66-6.6a.75.75 0 00-1.492-.149l-.66 6.6a.25.25 0 01-.249.225h-5.19a.25.25 0 01-.249-.225l-.66-6.6z'></path>"
+                            +"</svg>"
+                        +"</button>"
+                    +"</td>"
+                +"</tr>"
+            );
+        }else if(existente == 1){
+            Swal.fire(
+                {
+                    icon: 'warning',
+                    title: 'Producto ya agregado',
+                    text: 'Te recomendamos buscarlo en la tabla y sólo editar la cantidad asignada',
+                    showConfirmButton: false,
+                    timer: 1500
+                }
+            )
+        }
         document.getElementById("Producto").value = "";
         document.getElementById("Cantidad").value = "";
     }
